+++ Server/src/resolvers/Mutation/index.js
@@ -1,0 +1,11 @@
+export async function createUser(_, args, { db }, info) {
+  return db.mutation.createUser(args, info);
+}
+
+export async function updateUser(_, args, { db }, info) {
+  return db.mutation.updateUser(args, info);
+}
+
+export async function deleteUser(_, args, { db }, info) {
+  return db.mutation.deleteUser(args, info);
+}
