+++ Server/src/resolvers/Query/index.js
@@ -1,0 +1,7 @@
+export async function user(_, args, { db }, info) {
+  return db.query.user(args, info);
+}
+
+export async function users(_, args, { db }, info) {
+  return db.query.users(args, info);
+}
